Consertando Integração da Api
